backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
